+++ ubo_app/service_thread.py
@@ -27,7 +27,7 @@ from redux import (
 )
 
 from ubo_app.constants import (
-    DEBUG_MODE_TASKS,
+    DEBUG_TASKS,
     DISABLED_SERVICES,
     ENABLED_SERVICES,
     SERVICES_LOOP_GRACE_PERIOD,
@@ -405,14 +405,14 @@ class UboServiceThread(threading.Thread):
     ) -> asyncio.Handle:
         def task_wrapper(stack: str) -> None:
             task = self.loop.create_task(coroutine)
-            if DEBUG_MODE_TASKS:
+            if DEBUG_TASKS:
                 STACKS[task] = stack
             if callback:
                 callback(task)
 
         return self.loop.call_soon_threadsafe(
             task_wrapper,
-            ''.join(traceback.format_stack()[:-3]) if DEBUG_MODE_TASKS else '',
+            ''.join(traceback.format_stack()[:-3]) if DEBUG_TASKS else '',
         )
 
     async def shutdown(self: UboServiceThread) -> None: