+++ ubo_app/store/update_manager/reducer.py
@@ -53,7 +53,6 @@ def reducer(
             current_version=action.current_version,
             base_image_variant=action.base_image_variant,
             latest_version=action.latest_version,
-            serial_number=action.serial_number,
         )
         version_comparison = 1
         with contextlib.suppress(ValueError):